Backend/backendGlobal/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
Pillow==10.0.0  # Pour la gestion des images (futures fonctionnalités)
qrcode==7.4.2
